Change versioning scheme
Currently gnome shell integration for Chrome browser reports that new update
is avaiable when in fact the latest version is installed.
It is because of the difference in version scheme used by
the extension (semver) and extensions.gnome.org. When a new version is published,
then Gnome Extensions replaces the version number in metadata.json to a
single number.

Move from semantic versioning to single numbers, because of the above.