Improve loading process list
Use glibtop_get_proclist() method to load the list of processes instead
of current implementation which iterates over the files in /proc
directory.
In additon to that filter out system processes (like [kworker/2:0])
which is not killable by the user, therefore not useful to see it in the
list of processes.
Finally switch from Mainloop.idel_add() to GLib.idle_add() and set low
priority for asynchronous tasks such as loading process list.

Performance gain on order of magnitude by 2 (from 100ms to 6 ms).