Fix name collision in status area ID
When the extension registers the indicator widget with the Gnome Shell
top panel (aka. status area), it has to provide an ID (aka. role) which
must be unique across all extensions. It turned out that the current ID
"system-monitor" is used by other extension as well. This ID collision
causes all extensions sharing the same ID, except the first to fail to load.

This commit fixes that by using the extension's UUID which is unique by
definition.