Fix listing mounted filesystems
Currently the last two lines are removed from mount list, because it was
empty. However as kernel versions change, this changes as well. As a
result non-empty lines are removed, thus causing mount points not
showing up in the top mounted filesystems list.

This commit fixes that by removing empty lines properly.