Fix bug in Swap meter
The regexp does not produce a match when the line starts with
whitespace.
It's now fixed.

Certain processes are short lived, resulting the PID does not exist by
the time the code starts collecting swap memory staticstics. Also there
are special processes which cannot be swapped out from RAM. These does
not have a VmStat line in their statistics.
Now both errors are ignored. Futhermore such a PID won't be queried
again at next status update.

It not only make the functionality correct, but improves the performance
as well, because less PIDs need to be queried.