Eliminate shutter lag caused by Swap meter
Per process swap usage is calculated from /proc/<pid>/status files. To
list the top 3 processes in the dropdown panel, files are read for all active processes first.
In practice it means few 100 or even 1000 file reads must be done, in
order to find the top 3.
Measurements shows that for 250 PIDs, reading all stat files take about
40 ms. That's enough to hang up the main loop in Gnome Shell causing a noticable lag.

Even though async file reading is already utilised, but implemented
incorrectly in the extension code, making it ineffective.

This change combines aync Gio.File with JS Promises to unblock the main
loop. It does not necessary eliminate the lag completely, but
seems to be unnoticable at the moment.

Note: An optimal alternative could be using glibTOP as it's already
utilised in other scenarios. But it can give only cumulative swap info, not per
process info. Until that changes, the current file based approach
remains in place.