Fix UI layout
After each refresh period, the length of items in the dropdown panel can
vary. The ST widgets will automatically resized to fit tot he new itmes
length. That makes the UI hard to read, because the items are moving
vertically.

This fix freezes the width of meter widgets when the panel is opened and
unreeze when closed. This way the UI takes the size of the actual
content and fixes it until remain opened.