Feature/add gnome 45 compatibility (#96)
* Add Gnome Shell 45 compatibility

- Replace module import with the new syntax
- Replace log() calls with console.*() calls.
- Provide the main extension and preferences dialog as exported classes

* Make panel button work

* Fix UI layout

After each refresh period, the length of items in the dropdown panel can
vary. The ST widgets will automatically resized to fit tot he new itmes
length. That makes the UI hard to read, because the items are moving
vertically.

This fix freezes the width of meter widgets when the panel is opened and
unreeze when closed. This way the UI takes the size of the actual
content and fixes it until remain opened.

* Rewrite preferences widget

Make the extension preferences widget compatible with Gnome Shell 45.

* Fix CI build